Module conc_linked_list::concurrent_linked_list

Structs

ConcurrentLinkedList
Basically a concurrent stack with no pop method and with linear contains method, which is using locks However, contains does not guarantee that the item is not inside, since another thread may have added the element at the beginning

Node

Functions

arc_mut_new